# contracts/openapi/admin-v1.yaml
openapi: 3.0.3
info:
  title: qiuxs-blog Admin API
  version: 1.0.0
servers:
  - url: https://blog-admin.qiuxs.com
paths:
...
  /api/admin/v1/tags:
    get:
      operationId: listTags
      responses:
        '200':
          description: Tags in stable display order
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TagList'}
        '401': {$ref: '#/components/responses/ProblemResponse'}
        '503': {$ref: '#/components/responses/ProblemResponse'}
    post:
      operationId: createTag
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateTagRequest'}
      responses:
        '201':
          description: Created tag
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TagView'}
        '400': {$ref: '#/components/responses/ProblemResponse'}
        '401': {$ref: '#/components/responses/ProblemResponse'}
        '403': {$ref: '#/components/responses/ProblemResponse'}
        '409': {$ref: '#/components/responses/ProblemResponse'}
        '503': {$ref: '#/components/responses/ProblemResponse'}
  /api/admin/v1/tags/{tagId}:
    parameters:
      - {$ref: '#/components/parameters/TagId'}
    patch:
      operationId: renameTag
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RenameTagRequest'}
      responses:
        '200':
          description: Renamed tag
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TagView'}
        '400': {$ref: '#/components/responses/ProblemResponse'}
        '401': {$ref: '#/components/responses/ProblemResponse'}
        '403': {$ref: '#/components/responses/ProblemResponse'}
        '404': {$ref: '#/components/responses/ProblemResponse'}
        '409': {$ref: '#/components/responses/ProblemResponse'}
        '503': {$ref: '#/components/responses/ProblemResponse'}
...
  /api/admin/v1/settings/site:
    get:
      operationId: getSiteSettings
      responses:
        '200': {$ref: '#/components/responses/SiteSettingsResponse'}
        '401': {$ref: '#/components/responses/ProblemResponse'}
        '503': {$ref: '#/components/responses/ProblemResponse'}
    put:
      operationId: putSiteSettings
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PutSiteSettingsRequest'}
      responses:
        '200': {$ref: '#/components/responses/SiteSettingsResponse'}
        '400': {$ref: '#/components/responses/ProblemResponse'}
        '401': {$ref: '#/components/responses/ProblemResponse'}
        '403': {$ref: '#/components/responses/ProblemResponse'}
        '409': {$ref: '#/components/responses/ProblemResponse'}
        '422': {$ref: '#/components/responses/ProblemResponse'}
        '503': {$ref: '#/components/responses/ProblemResponse'}
  /api/admin/v1/settings/hotlink:
    get:
      operationId: getHotlinkSettings
      responses:
        '200':
          description: Current hotlink policy
          content:
            application/json:
              schema: {$ref: '#/components/schemas/HotlinkSettingsView'}
        '401': {$ref: '#/components/responses/ProblemResponse'}
        '503': {$ref: '#/components/responses/ProblemResponse'}
    put:
      operationId: putHotlinkSettings
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PutHotlinkSettingsRequest'}
      responses:
        '200':
          description: Replaced hotlink policy
          content:
            application/json:
              schema: {$ref: '#/components/schemas/HotlinkSettingsView'}
        '400': {$ref: '#/components/responses/ProblemResponse'}
        '401': {$ref: '#/components/responses/ProblemResponse'}
        '403': {$ref: '#/components/responses/ProblemResponse'}
        '422': {$ref: '#/components/responses/ProblemResponse'}
        '503': {$ref: '#/components/responses/ProblemResponse'}
components:
  parameters:
...
    TagId:
      name: tagId
      in: path
      required: true
      schema: {type: integer, format: int64, minimum: 1}
  responses:
    ProblemResponse:
      description: Request failed
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
...
    SiteSettingsResponse:
      description: Publishable site settings
      content:
        application/json:
          schema: {$ref: '#/components/schemas/SiteSettingsView'}
  schemas:
...
    TagView:
      type: object
      additionalProperties: false
      required: [id, name, slug, createdAt, updatedAt]
      properties:
        id: {type: integer, format: int64, minimum: 1}
        name: {type: string}
        slug: {type: string}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    TagList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/TagView'}
    CreateTagRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: {type: string, minLength: 1, maxLength: 64}
    RenameTagRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: {type: string, minLength: 1, maxLength: 64}
...
    SocialLink:
      type: object
      additionalProperties: false
      required: [label, url]
      properties:
        label: {type: string}
        url: {type: string, format: uri}
    SiteSettingsView:
      type: object
      additionalProperties: false
      required: [id, lockVersion, siteName, authorName, authorBio, homeStatus, aboutMd, socialLinks, seoDefaultTitle, seoDefaultDescription, seoDefaultImageMediaId, filingName, filingNumber, filingUrl, updatedAt]
      properties:
        id: {type: integer, format: int64, minimum: 1, nullable: true}
        lockVersion: {type: integer, format: int64, minimum: 0}
        siteName: {type: string}
        authorName: {type: string}
        authorBio: {type: string}
        homeStatus: {type: string}
        aboutMd: {type: string}
        socialLinks:
          type: array
          maxItems: 16
          items: {$ref: '#/components/schemas/SocialLink'}
        seoDefaultTitle: {type: string}
        seoDefaultDescription: {type: string}
        seoDefaultImageMediaId: {type: integer, format: int64, minimum: 1, nullable: true}
        filingName: {type: string}
        filingNumber: {type: string}
        filingUrl: {type: string, format: uri, readOnly: true}
        updatedAt: {type: string, format: date-time, nullable: true}
    PutSiteSettingsRequest:
      type: object
      additionalProperties: false
      required: [lockVersion, siteName, authorName, authorBio, homeStatus, aboutMd, socialLinks, seoDefaultTitle, seoDefaultDescription, seoDefaultImageMediaId, filingName, filingNumber]
      properties:
        lockVersion: {type: integer, format: int64, minimum: 0}
        siteName: {type: string}
        authorName: {type: string}
        authorBio: {type: string}
        homeStatus: {type: string}
        aboutMd: {type: string, maxLength: 2097152}
        socialLinks:
          type: array
          maxItems: 16
          items: {$ref: '#/components/schemas/SocialLink'}
        seoDefaultTitle: {type: string}
        seoDefaultDescription: {type: string}
        seoDefaultImageMediaId: {type: integer, format: int64, minimum: 1, nullable: true}
        filingName: {type: string}
        filingNumber: {type: string}
    HotlinkEntry:
      type: object
      additionalProperties: false
      required: [hostname, enabled]
      properties:
        hostname: {type: string}
        enabled: {type: boolean}
    HotlinkSettingsView:
      type: object
      additionalProperties: false
      required: [allowEmptyReferer, entries]
      properties:
        allowEmptyReferer: {type: boolean}
        entries:
          type: array
          items: {$ref: '#/components/schemas/HotlinkEntry'}
    PutHotlinkSettingsRequest:
      type: object
      additionalProperties: false
      required: [allowEmptyReferer, entries]
      properties:
        allowEmptyReferer: {type: boolean}
        entries:
          type: array
          items: {$ref: '#/components/schemas/HotlinkEntry'}
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, requestId]
      properties:
        type: {type: string, format: uri}
        title: {type: string}
        status: {type: integer}
        code: {type: string}
        requestId: {type: string}